+++ CJBCheatsMenu/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming release
 * The 'water all fields' cheat now works for indoor pots in any location.
+* You can now now override the farm and casino warps via `data/warp.json`.
 * Fixed health bonuses not applied when changing professions.
 * Fixed 'always auto-feed' cheat changing total hay incorrectly in some cases.
 * Fixed 'always auto-feed' cheat not counting animals who aren't in their home building if it's enabled after the day already started.